vkd3d: Handle fallback scenarios where a type mask does not exist.
If we request DEVICE | HOST_VISIBLE and the driver does not expose the
type at all, we'll hit fallback logic that assumes we actually failed to
allocate rather than failed to identify a candidate memory type.

Attempt fallbacks in the try functions as well if we did not identify a
suitable memory type at all.

Signed-off-by: Hans-Kristian Arntzen <post@arntzen-software.no>