vkd3d: Add PREALLOCATE_SRV_MIP_CLAMPS workaround for Halo Infinite
Halo Infinite currently has a race condition that can result in a GPU fault on
NVIDIA hardware. The root cause of this race condition is the application overwriting
existing SRV descriptors with a new mip clamp value asynchronously to the applications GPU
submission timeline. This is problematic for vkd3d because it requires creating a new VkImageView
to represent the SRV with the updated mip clamp value. If this SRV is being created after
work has been submitted that accesses it, the GPU may not see the correct descriptor state
possibly resulting in a fault.

This commit adds the VKD3D_CONFIG_FLAG_PREALLOCATE_SRV_MIP_CLAMPS configuration as a workaround to
the race condition. It works by preallocating all the VkImageViews for each possible mip clamp level
the application may attempt to create for the resource. Technically this is still a race condition
but the end result isn't a GPU fault if the race condition is lost. Instead the shader will just sample
from the texture with the previous SRVs mip clamp level rather than the updated one. These are
the same outcomes that are possible on RADV without this workaround.

Signed-off-by: Eric Sullivan <esullivan@nvidia.com>