vkd3d: Allocate private memory for host-visible images.
This extra memory is currently unused, but the plan is to move the
actual image resource to this private allocation and only expose
the linear staging buffer to the application.

However we'll need the complete implementation for that to work, but
the implementation will require the private memory and subresource
layout info to be present, so just leave some temporary TODOs for now.

Signed-off-by: Philip Rebohle <philip.rebohle@tu-dortmund.de>