vkd3d: Add global query for descriptor buffers.
Committing this here allows us to introduce new code little by little.
Also adds implementation for initializing the global descriptor buffers,
but do not call them yet. That will happen in the final commit.

Global descriptor buffers are used as a fallback when
application does not bind anything on its own.

We need to always have something bound for:
- Push descriptors with bufferlessPushDescriptors == VK_FALSE.
- Simplifies descriptor table update path where we can assume everything
  is valid.

Signed-off-by: Hans-Kristian Arntzen <post@arntzen-software.no>